Moved ILocatable to API.
