+++ src/main/java/forestry/core/multiblock/MultiblockControllerForestry.java
@@ -12,6 +12,8 @@ package forestry.core.multiblock;
 
 import javax.annotation.Nonnull;
 
+import forestry.api.core.ILocatable;
+
 import com.google.common.collect.HashMultiset;
 import com.google.common.collect.Multiset;
 import com.mojang.authlib.GameProfile;
@@ -24,7 +26,7 @@ import forestry.core.inventory.IInventoryAdapter;
 import forestry.core.owner.IOwnerHandler;
 import forestry.core.owner.IOwnedTile;
 import forestry.core.owner.OwnerHandler;
-import forestry.core.tiles.ILocatable;
+
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.ISidedInventory;
 import net.minecraft.item.ItemStack;
