+++ src/main/java/forestry/core/multiblock/MultiblockTileEntityForestry.java
@@ -13,6 +13,8 @@ package forestry.core.multiblock;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
+import forestry.api.core.ILocatable;
+
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.ISidedInventory;
 import net.minecraft.item.ItemStack;
@@ -22,7 +24,6 @@ import net.minecraft.util.text.ITextComponent;
 import net.minecraft.world.World;
 
 import com.mojang.authlib.GameProfile;
-
 import forestry.api.multiblock.IMultiblockLogic;
 import forestry.api.multiblock.MultiblockTileEntityBase;
 import forestry.core.config.Constants;
@@ -31,7 +32,6 @@ import forestry.core.gui.IGuiHandlerTile;
 import forestry.core.inventory.FakeInventoryAdapter;
 import forestry.core.inventory.IInventoryAdapter;
 import forestry.core.tiles.IFilterSlotDelegate;
-import forestry.core.tiles.ILocatable;
 import forestry.core.utils.PlayerUtil;
 
 public abstract class MultiblockTileEntityForestry<T extends IMultiblockLogic> extends MultiblockTileEntityBase<T> implements ISidedInventory, IFilterSlotDelegate, ILocatable, IGuiHandlerTile {
