+++ src/main/java/forestry/core/tiles/TileForestry.java
@@ -12,6 +12,9 @@ package forestry.core.tiles;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+
+import forestry.api.core.ILocatable;
+
 import java.io.IOException;
 import java.util.Random;
 
@@ -27,6 +30,7 @@ import forestry.core.network.IStreamable;
 import forestry.core.network.packets.PacketTileStream;
 import forestry.core.proxy.Proxies;
 import forestry.core.utils.NBTUtilForestry;
+
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.ISidedInventory;
 import net.minecraft.item.ItemStack;
@@ -40,6 +44,7 @@ import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
+
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.wrapper.SidedInvWrapper;
